Add andino_firmware include directory and move HAL interfaces
