+++ andino_firmware/platformio.ini
@@ -14,6 +14,7 @@ default_envs = uno, nanoatmega328
 ; Base configuration for build tools.
 [base_build]
 build_flags =
+  -I include
   -Wall -Wextra
 
 ; Base configuration for Atmel AVR based Arduino boards.
