+++ andino_firmware/src/digital_out_arduino.h
@@ -29,7 +29,7 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 
-#include "digital_out.h"
+#include "andino/hal/digital_out.h"
 
 namespace andino {
