+++ andino_firmware/src/motor.h
@@ -64,8 +64,8 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 
-#include "digital_out.h"
-#include "pwm_out.h"
+#include "andino/hal/digital_out.h"
+#include "andino/hal/pwm_out.h"
 
 namespace andino {
