+++ andino_firmware/src/pwm_out_arduino.h
@@ -29,7 +29,7 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 
-#include "pwm_out.h"
+#include "andino/hal/pwm_out.h"
 
 namespace andino {
