+++ andino_firmware/src/shell.h
@@ -31,7 +31,7 @@
 
 #include <stddef.h>
 
-#include "serial_stream.h"
+#include "andino/hal/serial_stream.h"
 
 namespace andino {
